Use 64 bit integers for some internal counters, not 32 bits.
32 bits overflow at only 2 147 483 648 (2^31), which is inappropriate
for counting RPCs or events related to RPCs, since after a few days of
uptime a server can easily serve that many requests...  Doh!

+++ src/tsd/ConnectionManager.java
@@ -13,7 +13,7 @@
 package net.opentsdb.tsd;
 
 import java.nio.channels.ClosedChannelException;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -35,10 +35,8 @@ final class ConnectionManager extends SimpleChannelHandler {
 
   private static final Logger LOG = LoggerFactory.getLogger(ConnectionManager.class);
 
-  private static final AtomicInteger connections_established
-    = new AtomicInteger();
-  private static final AtomicInteger exceptions_caught
-    = new AtomicInteger();
+  private static final AtomicLong connections_established = new AtomicLong();
+  private static final AtomicLong exceptions_caught = new AtomicLong();
 
   private static final DefaultChannelGroup channels =
     new DefaultChannelGroup("all-channels");
